Add further routes to Posts and Users controllers
