+++ src/services/http/_fixtures.ts
@@ -1,0 +1,29 @@
+export const post = {
+  type: 'POST',
+  subtype: '',
+  messageId:
+    '0x3E1E26f055Cd29053D44Fc65aa1FCa216DedFceb/67e929dd44b631b80e22bd0f94b1b75812e0159117d24bbfba592c9340804fa6',
+  hash: '67e929dd44b631b80e22bd0f94b1b75812e0159117d24bbfba592c9340804fa6',
+  createdAt: '1648260542270',
+  payload: { topic: '', title: '', content: 'auti.sm is a pwa ', reference: '', attachment: '' },
+  meta: {
+    replyCount: 2,
+    likeCount: 2,
+    repostCount: 0,
+    liked: null,
+    reposted: null,
+    blocked: null,
+    interepProvider: null,
+    interepGroup: null,
+    rootId:
+      '0x3E1E26f055Cd29053D44Fc65aa1FCa216DedFceb/67e929dd44b631b80e22bd0f94b1b75812e0159117d24bbfba592c9340804fa6',
+    moderation: 'THREAD_HIDE_BLOCK',
+    modblockedctx: null,
+    modfollowedctx: null,
+    modmentionedctx: null,
+    modLikedPost: null,
+    modBlockedPost: null,
+    modBlockedUser: null,
+    modFollowerUser: null,
+  },
+};
