+++ src/services/http/http.test.ts
@@ -3,157 +3,7 @@ import tape from 'tape';
 import sinon from 'sinon';
 import HttpService from '.';
 import { newRequest, newResponse, stubCall, stubFetch } from '../../util/testUtils';
-
-const post = {
-  type: 'POST',
-  subtype: '',
-  messageId:
-    '0x3E1E26f055Cd29053D44Fc65aa1FCa216DedFceb/67e929dd44b631b80e22bd0f94b1b75812e0159117d24bbfba592c9340804fa6',
-  hash: '67e929dd44b631b80e22bd0f94b1b75812e0159117d24bbfba592c9340804fa6',
-  createdAt: '1648260542270',
-  payload: { topic: '', title: '', content: 'auti.sm is a pwa ', reference: '', attachment: '' },
-  meta: {
-    replyCount: 2,
-    likeCount: 2,
-    repostCount: 0,
-    liked: null,
-    reposted: null,
-    blocked: null,
-    interepProvider: null,
-    interepGroup: null,
-    rootId:
-      '0x3E1E26f055Cd29053D44Fc65aa1FCa216DedFceb/67e929dd44b631b80e22bd0f94b1b75812e0159117d24bbfba592c9340804fa6',
-    moderation: 'THREAD_HIDE_BLOCK',
-    modblockedctx: null,
-    modfollowedctx: null,
-    modmentionedctx: null,
-    modLikedPost: null,
-    modBlockedPost: null,
-    modBlockedUser: null,
-    modFollowerUser: null,
-  },
-};
-
-tape('HTTPService - handleGetReplies', async t => {
-  const http = new HttpService();
-  const [call, stubs] = stubCall(http);
-  const req = newRequest(null, null, {
-    parent: '0xparenthash/67e929dd44b631b80e22bd0f94b1b75812e0159117d24bbfba592c9340804fa6',
-  });
-  const res = newResponse();
-
-  stubs.posts.findOne.returns(
-    Promise.resolve({
-      subtype: '',
-      payload: {},
-    })
-  );
-
-  stubs.posts.findAllReplies.returns(Promise.resolve([{ hash: '0xposthash' }]));
-
-  await http.handleGetReplies(req, res);
-
-  t.deepEqual(
-    stubs.posts.findAllReplies.args[0],
-    [
-      '0xparenthash/67e929dd44b631b80e22bd0f94b1b75812e0159117d24bbfba592c9340804fa6',
-      undefined,
-      0,
-      10,
-      'ASC',
-      undefined,
-      false,
-    ],
-    'should find all replies'
-  );
-
-  t.deepEqual(
-    res.send.args[0],
-    [
-      {
-        payload: [
-          {
-            hash: '0xposthash',
-          },
-        ],
-        error: undefined,
-      },
-    ],
-    'should return replies'
-  );
-
-  stubs.posts.findOne.returns(
-    Promise.resolve({
-      subtype: 'M_POST',
-      payload: {
-        topic: 'https://twitter.com/0xTsukino/status/1465780936314740736',
-      },
-    })
-  );
-
-  stubs.posts.findAllReplies.returns(Promise.resolve([{ hash: '0xposthash' }]));
-  stubs.posts.findLastTweetInConversation.returns(Promise.resolve({ hash: '0xtweethash' }));
-
-  const fetch = stubFetch();
-  fetch.reset();
-  fetch.returns(
-    Promise.resolve({
-      json: async () => ({
-        data: [
-          {
-            author_id: 'yagami',
-            created_at: '1234',
-            coversation_id: 'threadid',
-            in_reply_to_user_id: 'replyid',
-            text: 'hello!',
-            id: '',
-            referenced_tweets: [],
-          },
-        ],
-      }),
-    })
-  );
-  await http.handleGetReplies(req, res);
-
-  t.deepEqual(
-    fetch.args[0],
-    [
-      'https://api.twitter.com/2/tweets/search/recent?query=conversation_id:1465780936314740736&since_id=0xtweethash&max_results=100&expansions=author_id,in_reply_to_user_id&tweet.fields=referenced_tweets,in_reply_to_user_id,author_id,created_at,conversation_id&user.fields=name,username',
-      {
-        method: 'GET',
-        headers: {
-          Authorization: 'Bearer twBearerToken',
-        },
-      },
-    ],
-    'should fetch twitter replies'
-  );
-
-  t.deepEqual(
-    stubs.posts.createTwitterPosts.args[0],
-    [
-      [
-        {
-          messageId: '',
-          hash: '',
-          creator: 'yagami',
-          type: '@TWEET@',
-          subtype: '',
-          createdAt: -23225875200000,
-          topic: '',
-          title: '',
-          content: 'hello!',
-          reference: '1465780936314740736',
-          attachment: '',
-        },
-      ],
-    ],
-    'should create twitter replies'
-  );
-
-  fetch.reset();
-  t.end();
-});
+import { post } from './_fixtures';
 
 tape('HTTPService - handleGetPostsByTag', async t => {
   const http = new HttpService();
@@ -193,56 +43,6 @@ tape('HTTPService - handleGetTags', async t => {
     [{ payload: [{ tagName: '#test' }], error: undefined }],
     'should return all tags'
   );
-
-  t.end();
-});
-
-tape('HTTPService - handleGetUserReplies', async t => {
-  const http = new HttpService();
-  const [call, stubs] = stubCall(http);
-  const req = newRequest({
-    creator: '0xmyuser',
-  });
-  const res = newResponse();
-
-  stubs.posts.findAllRepliesFromCreator.returns(Promise.resolve([post]));
-
-  await http.handleGetUserReplies(req, res);
-
-  t.deepEqual(
-    stubs.posts.findAllRepliesFromCreator.args[0],
-    ['0xmyuser', undefined, 0, 10],
-    'should find all replies'
-  );
-
-  t.deepEqual(
-    res.send.args[0],
-    [{ payload: [post], error: undefined }],
-    'should return all replies'
-  );
-
-  t.end();
-});
-
-tape('HTTPService - handleGetUserLikes', async t => {
-  const http = new HttpService();
-  const [call, stubs] = stubCall(http);
-  const req = newRequest({
-    creator: '0xmyuser',
-  });
-  const res = newResponse();
-
-  stubs.posts.findAllLikedPostsByCreator.returns(Promise.resolve([post]));
-
-  await http.handleGetUserLikes(req, res);
-
-  t.deepEqual(
-    stubs.posts.findAllLikedPostsByCreator.args[0],
-    ['0xmyuser', undefined, 0, 10],
-    'should find all likes'
-  );
-
-  t.deepEqual(res.send.args[0], [{ payload: [post], error: undefined }], 'should return all likes');
 
   t.end();
 });
@@ -404,36 +204,6 @@ tape.skip('HTTPService - get preview', async t => {
     'should return correct result from interep'
   );
 
-  t.end();
-});
-
-tape('HttpService - get followers per user', async t => {
-  // const route = '/v1/post/:hash/likes';
-  const http = new HttpService();
-  const [, stubs] = stubCall(http);
-  const res = newResponse();
-  const followers = ['0xfoo', '0xbar'];
-
-  stubs.connections.findAllFollowersByName.returns(Promise.resolve(followers));
-
-  await http.handleGetUserFollowers(newRequest({ user: '0xr1oga' }, null, null), res);
-
-  t.deepEqual(res.send.args[0][0].payload, followers, 'should be equal');
-  t.end();
-});
-
-tape('HttpService - get followings per user', async t => {
-  // const route = '/v1/post/:hash/likes';
-  const http = new HttpService();
-  const [, stubs] = stubCall(http);
-  const res = newResponse();
-  const followings = ['0xfoo', '0xbar'];
-
-  stubs.connections.findAllFollowingsByCreator.returns(Promise.resolve(followings));
-
-  await http.handleGetUserFollowings(newRequest({ user: '0xr1oga' }, null, null), res);
-
-  t.deepEqual(res.send.args[0][0].payload, followings, 'should be equal');
   t.end();
 });
